Pull common entity page into wrapper
