+++ context/app/static/js/components/entitySearch/EntitySearchWrapper/index.js
@@ -1,0 +1,3 @@
+import EntitySearchWrapper from './EntitySearchWrapper';
+
+export default EntitySearchWrapper;
